+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 * Remove Clndr dependency
 * Remove Moment dependency
+* Admin settings for each job filter
+* Frontend filters
+* Frontend loading effect
 
 ### 1.0.0: July 21th, 2020
 * Plugin boilerplate